+++ game/dota_addons/barebones/scripts/vscripts/events.lua
@@ -332,3 +332,12 @@ function GameMode:OnNPCGoalReached(keys)
   local nextGoalEntity = EntIndexToHScript(keys.next_goal_entindex)
   local npc = EntIndexToHScript(keys.npc_entindex)
 end
+
+-- This function is called whenever any player sends a chat message to team or All
+function GameMode:OnPlayerChat(keys)
+  local teamonly = keys.teamonly
+  local userID = keys.userid
+  local playerID = self.vUserIds[userID]:GetPlayerID()
+
+  local text = keys.text
+end