+++ game/dota_addons/barebones/scripts/vscripts/internal/gamemode.lua
@@ -88,6 +88,8 @@ function GameMode:_InitGameMode()
   ListenToGameEvent("dota_tower_kill", Dynamic_Wrap(GameMode, 'OnTowerKill'), self)
   ListenToGameEvent("dota_player_selected_custom_team", Dynamic_Wrap(GameMode, 'OnPlayerSelectedCustomTeam'), self)
   ListenToGameEvent("dota_npc_goal_reached", Dynamic_Wrap(GameMode, 'OnNPCGoalReached'), self)
+
+  ListenToGameEvent("player_chat", Dynamic_Wrap(GameMode, 'OnPlayerChat'), self)
   
   --ListenToGameEvent("dota_tutorial_shop_toggled", Dynamic_Wrap(GameMode, 'OnShopToggled'), self)
 
@@ -117,6 +119,7 @@ function GameMode:_InitGameMode()
 
   -- Initialized tables for tracking state
   self.bSeenWaitForPlayers = false
+  self.vUserIds = {}
 
   DebugPrint('[BAREBONES] Done loading Barebones gamemode!\n\n')
 end